Show account backups card on dashboard unless user has Managed
